Fix: Unintended remove of AppInfo

+++ app/src/main/java/com/machiav3lli/backup/services/PackageUnInstalledReceiver.kt
@@ -51,7 +51,8 @@ class PackageUnInstalledReceiver : BroadcastReceiver() {
                 }
                 Intent.ACTION_PACKAGE_REMOVED -> {
                     GlobalScope.launch(Dispatchers.IO) {
-                        db.appInfoDao.deleteAllOf(packageName)
+                        if (db.backupDao.get(packageName).isEmpty())
+                            db.appInfoDao.deleteAllOf(packageName)
                     }
                 }
             }
